openai_mock will now record even if code fails and allows starting with replay and continue with recording

+++ tests/utils.py
@@ -4,37 +4,47 @@ import re
 import os
 
 from contextlib import contextmanager
+from typing import Optional
 
 from scientistgpt import Message, Conversation
 from scientistgpt.conversation import conversation
 
 
 @contextmanager
-def mock_openai(responses):
-    index = -1
-
+def mock_openai(responses, record_more_from_openai_if_needed=False, fail_if_not_all_responses_used=True):
     def mock_chatgpt_response(messages: list[Message]):
-        nonlocal index
-        index += 1
-        if isinstance(responses[index], Exception):
-            raise responses[index]
-        return responses[index]
+        if not responses:
+            if record_more_from_openai_if_needed:
+                response = original_method(messages)
+                new_responses.append(response)
+            else:
+                raise AssertionError('No more responses were mocked')
+        else:
+            response = responses.pop(0)
+        if isinstance(response, Exception):
+            raise response
+        return response
 
+    new_responses = []
+    new_responses_and_exceptions = {'new_responses': new_responses, 'exception': None}
     original_method = conversation._get_chatgpt_response
     conversation._get_chatgpt_response = mock_chatgpt_response
-    yield
-    conversation._get_chatgpt_response = original_method
+    try:
+        yield new_responses_and_exceptions
+    except Exception as e:
+        new_responses_and_exceptions['exception'] = e
+    finally:
+        conversation._get_chatgpt_response = original_method
+    if responses and fail_if_not_all_responses_used:
+        raise AssertionError(f'Not all responses were used: {responses}')
 
 
 @contextmanager
 def record_openai():
-    index = -1
     original_method = conversation._get_chatgpt_response
     responses = []
 
     def record_chatgpt_response(messages: list[Message]):
-        nonlocal index
-        index += 1
         try:
             response = original_method(messages)
         except Exception as e:
@@ -63,22 +73,32 @@ def record_or_replay_openai(func):
 
         test_file_path = os.path.abspath(inspect.getfile(func))
         test_dir = os.path.dirname(test_file_path)
-        test_file = os.path.basename(test_file_path)
         file_path = os.path.join(test_dir, 'openai_responses', func.__name__ + '.txt')
 
+        # Create the directory and file if not exist
         if not os.path.exists(os.path.dirname(file_path)):
             os.makedirs(os.path.dirname(file_path))
 
-        try:
-            with open(file_path, 'r') as f:
-                responses = re.findall(Conversation.SAVE_START + "(.*?)" + Conversation.SAVE_END, f.read(), re.DOTALL)
-        except FileNotFoundError:
-            with record_openai() as responses:
-                func(*args, **kwargs)
+        if not os.path.isfile(file_path):
             with open(file_path, 'w') as f:
-                for response in responses:
+                f.write('')
+
+        # get previous responses (or empty list)
+        with open(file_path, 'r') as f:
+            responses = re.findall(Conversation.SAVE_START + "(.*?)" + Conversation.SAVE_END, f.read(), re.DOTALL)
+
+        # run the test with the previous responses and record new responses
+        with mock_openai(responses, record_more_from_openai_if_needed=True) as new_responses_and_exceptions:
+            func(*args, **kwargs)
+
+        # add new responses to the file
+        if new_responses_and_exceptions['new_responses']:
+            with open(file_path, 'a') as f:
+                for response in new_responses_and_exceptions['new_responses']:
                     f.write(f'{Conversation.SAVE_START}{response}{Conversation.SAVE_END}\n')
-        else:
-            with mock_openai(responses):
-                func(*args, **kwargs)
+
+        # raise exception if there was one
+        if new_responses_and_exceptions['exception']:
+            raise new_responses_and_exceptions['exception']
     return wrapper
+
